refactored window
